requirements.txt
flask
websocket-client
pandas
gunicorn